feat(backend): setup Car entity and repository with DB connection
- Updated Car entity with @Entity annotation and Long id
- Adjusted application-dev.properties with correct database name
- Refactored findById method to throw exception if car not found
- Verified save and findById endpoints in Postman

+++ backend/src/main/java/com/cars/backend/controller/CarController.java
@@ -14,7 +14,7 @@ public class CarController {
     private CarService carService;
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> findById(@PathVariable int id) {
+    public ResponseEntity<?> findById(@PathVariable Long id) {
         try {
             Car car = this.carService.findById(id);
             return ResponseEntity.ok(car);
